add corrections on fechamento layout, account to client and solved date bug

+++ migrations/20220131142142-create-fechamento.js
@@ -60,7 +60,7 @@ module.exports = {
         type: Sequelize.STRING
       },
       preco: {
-        type: Sequelize.FLOAT
+        type: Sequelize.DECIMAL(10,2)
       },
       quantidade: {
         type: Sequelize.FLOAT
